Multiple Indicator Cluster Survey (MICS) 2014 Child Questionnaire Data for Punjab, Pakistan

MICS2014Ch data of Child questionnaire, MICS Punjab, 2014.

data(MICS2014Ch)

Format

A data frame with 31083 rows and 270 variables:

HH1

Cluster number

HH2

Household number

LN

Line number

UF1

Cluster number

UF2

Household number

UF4

Child's line number

UF6

Mother / Caretaker's line number

UF7

Interviewer number

UF8D

Day of interview

UF8M

Month of interview

UF8Y

Year of interview

UF9

Result of interview for children under 5

UF10

Field editor

UF11

Main data entry clerk

UF12H

Start of interview - Hour

UF12M

Start of interview - Minutes

UF13H

End of interview - Hour

UF13M

End of interview - Minutes

AG1D

Day of birth of child

AG1M

Month of birth of child

AG1Y

Year of birth of child

AG2

Age of child

BR1

Birth certificate

BR2

Birth registered

BR3

Know how to register birth

EC1

Number of children's books or picture books for child

EC2A

Homemade toys

EC2B

Toys from shops

EC2C

Household objects or outside objects

EC3A

In past week, days left alone for more than 1 hour

EC3B

In past week, days left with other child for more than 1 hour

EC5

Attends early childhood education programme

EC7AA

Read books - Mother

EC7AB

Read books - Father

EC7AX

Read books - Other

EC7AY

Read books - No one

EC7BA

Tell stories - Mother

EC7BB

Tell stories - Father

EC7BX

Tell stories - Other

EC7BY

Tell stories - No one

EC7CA

Sang songs - Mother

EC7CB

Sang songs - Father

EC7CX

Sang songs - Other

EC7CY

Sang songs - No one

EC7DA

Took outside - Mother

EC7DB

Took outside - Father

EC7DX

Took outside - Other

EC7DY

Took outside - No one

EC7EA

Played with - Mother

EC7EB

Played with - Father

EC7EX

Played with - Other

EC7EY

Played with - No one

EC7FA

Named/counted - Mother

EC7FB

Named/counted - Father

EC7FX

Named/counted - Other

EC7FY

Named/counted - No one

EC8

Child identifies at least ten letters of the alphabet

EC9

Child reads at least four simple, popular words

EC10

Child knows name and recognizes symbol of all numbers from 1-10

EC11

Child able to pick up small object with 2 fingers

EC12

Child sometimes too sick to play

EC13

Child follows simple directions

EC14

Child able to do something independently

EC15

Child gets along well with other children

EC16

Child kicks, bites or hits other children or adults

EC17

Child gets distracted easily

BD2

Child ever been breastfed

BD3

Child still being breastfed

BD4

Child drank anything else from the bottle with a nipple yesterday

BD5

Child drank ORS yesterday

BD6

Child drank or ate vitamin or mineral supplements yesterday

BD7A

Child drank plain water yesterday

BD7B

Child drank juice or juice drinks yesterday

BD7C

Child drank clear soup (any type) yesterday

BD7D

Child drank milk such as tinned, powdered, fresh animal milk or Cured shake (Lasi) yesterday

BD7DN

Times child drank milk

BD7E

Child drank infant formula yesterday

BD7EN

Times child drank infant formula

BD7F

Child drank any other liquid yesterday

BD8A

Child drank or ate yogurt yesterday

BD8AN

Times drank or ate yogurt

BD8B

Child ate cerelac or any fortified baby food

BD8C

Child ate foods made from grains (Dalia etc)

BD8D

Child ate pumpkin, carrots, squash etc.

BD8E

Child ate white potatoes, white yams, manioc, turnip, spinach, green beans etc.

BD8F

Child ate green leafy vegetables

BD8G

Child ate ripe mangoes, papayas

BD8H

Child ate other fruits or vegetables

BD8I

Child ate liver, kidney, heart or other organ meat

BD8J

Child ate meat, such as beef, pork, lamb, goat, chicken, duck

BD8K

Child ate eggs

BD8L

Child ate fresh or dried fish or shellfish

BD8M

Child ate beans, lentils or nuts

BD8N

Child ate cheese or other food made from milk

BD8O

Child ate other solid, semi-solid or soft food yesterday

BD11

Times child ate solid or semi-solid food

IM1

Vaccination card for child

IM2

Ever had vaccination card

IM3BD

Day of BCG immunization

IM3BM

Month of BCG immunization

IM3BY

Year of BCG immunization

IM3P0D

Day of Polio at birth immunization

IM3P0M

Month of Polio at birth immunization

IM3P0Y

Year of Polio at birth immunization

IM3P1D

Day of Polio1 immunization

IM3P1M

Month of Polio1 immunization

IM3P1Y

Year of Polio1 immunization

IM3P2D

Day of Polio2 immunization

IM3P2M

Month of Polio2 immunization

IM3P2Y

Year of Polio2 immunization

IM3P3D

Day of Polio3 immunization

IM3P3M

Month of Polio3 immunization

IM3P3Y

Year of Polio3 immunization

IM3T1D

Day of PENTA1 immunization

IM3T1M

Month of PENTA1 immunization

IM3T1Y

Year of PENTA1 immunization

IM3T2D

Day of PENTA2 immunization

IM3T2M

Month of PENTA2 immunization

IM3T2Y

Year of PENTA2 immunization

IM3T3D

Day of PENTA3 immunization

IM3T3M

Month of PENTA3 immunization

IM3T3Y

Year of PENTA3 immunization

IM3M1D

Day measles1 or MMR immunization

IM3M1M

Month Measles1 or MMR immunization

IM3M1Y

Year of Measles1 or MMR immunization

IM3M2D

Day Measles2 or MMR immunization

IM3M2M

Month Measles2 or MMR immunization

IM3M2Y

Year Measles2 or MMR immunization

IM5

Child received any other vaccinations

IM6

Child ever received any vaccinations

IM7

Child ever given BCG vaccination

IM8

Child ever given Polio vaccination

IM9

Polio first given just after birth or later

IM10

Times child given Polio vaccination

IM11

Child ever given DPT vaccination

IM12

Times child given DPT vaccination

IM15A

Child ever given Pentavalent vaccination

IM15B

Times child given Pentavalent vaccination

IM16

Child ever given Measles or MMR vaccination

IM19A

Child participated in: Anti-Polio campaign - NID (-------------

IM19B

Child participated in: Mother and Child week (---------------)

IM19C

Child participated in: Child Health Day (---------------------)

VS1

Child received a Vitamin A dose within the last 6 months

CA1

Child had diarrhoea in last 2 weeks

CA2

Child drank less or more during illness

CA3

Child ate less or more during illness

CA3A

Sought advice or treatment for the illness from any source

CA3BA

Place sought care: (public sector) Government hospital

CA3BB

Place sought care: (public sector) Government health center

CA3BC

Place sought care: (public sector) Government health post/dispensary

CA3BD

Place sought care: (public sector) Lady health worker (LHW)

CA3BE

Place sought care: (public sector) Mobile / Outreach clinic

CA3BH

Place sought care: Other public

CA3BI

Place sought care: Private hospital / clinic

CA3BJ

Place sought care: Private physician

CA3BK

Place sought care: Private pharmacy

CA3BL

Place sought care: Mobile clinic

CA3BO

Place sought care: Other private medical

CA3BP

Place sought care: Relative / Friend

CA3BQ

Place sought care: Shop

CA3BR

Place sought care: Traditional practitioner

CA3BX

Place sought care: Other

CA4A

Drank fluid made from special packet (ORS)

CA4B

Pre-packaged ORS fluid for diarrhoea

CA4BB

Place where ORS was obtained

CA4CA

Child given zinc tablets

CA4CB

Child given zinc syrup

CA4E

Place where zinc was obtained

CA4FA

Child given home made fluid (boiled water, sugar and salt)

CA4FB

Child given home made fluid (Others - specify)

CA5

Anything else given to treat the diarrhoea

CA6A

Other treatment (pill or syrup): Antibiotic

CA6B

Other treatment (pill or syrup): Antimotility

CA6G

Other treatment (pill or syrup): Other (not antibiotic, antimotility or zinc)

CA6H

Other treatment (pill or syrup): Unknown

CA6L

Other treatment (injection): Antibiotic

CA6M

Other treatment (injection): Non-antibiotic

CA6N

Other treatment (injection): Unknown

CA6O

Other treatment: Intravenous

CA6Q

Other treatment: Home remedy/Herbal medicine

CA6X

Other treatment: Other

CA6AA

Child ill with fever in last 2 weeks

CA6BB

During fever blood was taken from his/her finger of heel

CA7

Child ill with cough in last 2 weeks

CA8

Difficulty breathing during illness with cough

CA9

Symptoms due to problem in chest or blocked nose

CA10

Sought advice or treatment for illness

CA11A

Place sought care: (public sector) Government hospital

CA11B

Place sought care: (public sector) Government health center

CA11C

Place sought care: (public sector) Government health post/dispensary

CA11D

Place sought care: (public sector) Lady health worker (LHW)

CA11E

Place sought care: (public sector) Mobile / Outreach clinic

CA11H

Place sought care: Other public

CA11I

Place sought care: Private hospital / clinic

CA11J

Place sought care: Private physician

CA11K

Place sought care: Private pharmacy

CA11L

Place sought care: Mobile clinic

CA11O

Place sought care: Other private medical

CA11P

Place sought care: Relative / Friend

CA11Q

Place sought care: Shop

CA11R

Place sought care: Traditional practitioner

CA11X

Place sought care: Other

CA12

Given medicine to treat this illness

CA13A

Medicine: SP / Fansidar

CA13B

Medicine: Chloroquine

CA13C

Medicine: Amodiaquine

CA13D

Medicine: Quinine

CA13E

Medicine: Combiniation with Artemisinin

CA13H

Medicine: Other anti-malarial

CA13I

Medicine: Antibiotic pill / syrup

CA13J

Medicine: Antibiotic injection

CA13P

Medicine: Paracetamol / Panadol / Acetaminophen

CA13Q

Medicine: Aspirin

CA13R

Medicine: Ibupropfen

CA13X

Medicine: Other

CA13Z

Medicine: DK

CA13BB

Place where antibiotics were obtained

CA13DD

Place where anti-malarials were obtained

CA13EE

When was anti-malarial first taken

CA15

What was done to dispose of the stools

AN1

Measurer's identification code

AN2

Result of height / length and weight measurement

AN3

Child's weight (kilograms)

AN3A

Child undressed to the minimum

AN4

Child's length or height (centimetres)

AN4A

Child measured lying or standing

HAP

Height for age percentile NCHS

HAZ

Height for age z-score NCHS

HAM

Height for age percent of reference median NCHS

WAP

Weight for age percentile NCHS

WAZ

Weight for age z-score NCHS

WAM

Weight for age percent of reference median NCHS

WHP

Weight for height percentile NCHS

WHZ

Weight for height z-score NCHS

WHM

Weight for height percent of reference median NCHS

FLAG

Flag for anthropometric indicators

BMI

Body Mass Index WHO

ZBMI

Body Mass Index z-score WHO

HAZ2

Height for age z-score WHO

WAZ2

Weight for age z-score WHO

WHZ2

Weight for height z-score WHO

HAZFLAG

Height for age flag WHO

WAZFLAG

Weight for age flag WHO

WHZFLAG

Weight for height flag WHO

BMIFLAG

BMI flag WHO

WHZNOAGE

Weight for height - Age flag WHO

HH6

Area (Original)

HH7

District (Original)

HL4

Sex

ED4A

Highest level of education attended

ED4B

Highest grade completed at that level

CDOI

Date of interview child (CMC)

CDOB

Date of birth of child (CMC)

CAGE

Age (months)

CAGE.6

Age

CAGE.11

Age

CAGED

Age in days

melevel

Mother's education

division

Division

hh7r

District

hh6r

Area

suburban

stratum

Straum (Original)

wscore

Combined wealth score

windex5

Wealth index quintile

wscoreu

Urban wealth score

windex5u

Urban wealth index quintile

wscorer

Rural wealth score

windex5r

Rural wealth index quintile

chweight

Children's sample weight

References

UNICEF MICS (http://www.mics.unicef.org/surveys).

Examples

data(MICS2014Ch)